Add error handling via result objects

+++ src/CounterService/src/Application/Counters/DecrementCounter.cs
@@ -5,7 +5,7 @@ namespace DistributedCounter.CounterService.Application.Counters;
 
 public static class DecrementCounter
 {
-    public record Command(Guid CounterId, uint Amount) : IRequest<Response>;
+    public record Command(Guid CounterId, ulong Amount) : IRequest<Response>;
 
     public record Response
     {
