+++ src/CounterService/src/Domain/Counters/Events/CounterIncrementedEvent.cs
@@ -1,3 +1,3 @@
 ﻿namespace DistributedCounter.CounterService.Domain.Counters.Events;
 
-public record CounterIncrementedEvent(uint Amount);
+public record CounterIncrementedEvent(ulong Amount);
